third_party/blink/web_tests/external/wpt/css/css-properties-values-api/registered-property-cssom-expected.txt

This is a testharness.js-based test.
FAIL Initially unregistered values from stylesheet assert_equals: expected "10px" but got " 10px"
PASS CSSOM setters function as expected for unregistered properties
PASS CSS.registerProperty
PASS Formerly valid values are still readable from inline styles but are computed as the unset value
PASS Values not matching the registered type can still be set
PASS Values can be removed from inline styles
PASS Stylesheets can be modified by CSSOM
PASS Valid values can be set on inline styles
Harness: the test ran to completion.
